Discord Bot/requirements.txt
discord.py==1.7.3
aiohttp==3.8.1
beautifulsoup4==4.10.0
